Objects and rendering

+++ GDENG03_DX/pixelShader.hlsl
@@ -15,5 +15,5 @@ cbuffer constant: register(b0)
 
 float4 main(PS_INPUT input) : SV_TARGET
 {
-	return float4(lerp(input.color, input.color1, (float)((sin((float)(m_time / (float)500.0f)) + 1.0f) / 2.0f)),1.0f);
+	return float4(lerp(input.color, input.color1, (sin(m_time/500.0f) + 1.0f) / 2.0f),1.0f);
 }
